knowledge-graph-ingest/requirements.txt
python-docx==1.1.2
neo4j==5.26.0
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.6
pymupdf4llm==0.0.17